[FIX] Company adress & synchronization
Before this commit, company onchanges did not work orrectly as there was an additional old style onchange in the view
Additionally the whole calculation of city_id was not correctly in line with the computed address fields which synchronize with the "contact partner" found on the related partner. In order to solve this more elegantly part of the v10 code was backported, which does not work with traditional function fields any more.

Signed-off-by: David Arnold <dar@devco.co>